Test a dist with semantic versioning and a dev version (#11)
This path overrides CPAN::Meta::Converter::_clean_version
so that the string in the the `provides` is passed on 
literally instead of interpreted. If you use something
like v1.0.0_02, you get the same string back. This is 
despite the wisdom of using that (see the GitHub thread).

This is related to Perl-Toolchain-Gang/CPAN-Meta#138 which 
makes the same change (but has not been merged).

Note that all of this depends on the `provides` section
having the right value. If you generate META.json or 
META.yaml with the current CPAN::Meta, it may modify
the version value that it puts under `provides` even though
the top-level `version` is unmolested. That is, this
fix is half the fix and the rest depends on the distro.

If this really bugs you, you'll have to hand-edit
the META.* until the other side of the fix is in.